internal/indexer/requirements.txt
kafka-python==2.0.2
beautifulsoup4==4.12.3
lxml==5.2.2
elasticsearch==8.13.0
PyYAML==6.0.1
